pyproject.toml
[project]
name = "par_gpt"
dynamic = [
    "version",
]
description = "Par Gpt"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Paul Robello", email = "probello@gmail.com" },
]
maintainers = [
    { name = "Paul Robello", email = "probello@gmail.com" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Other Audience",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = [
    "ai",
    "langchain",
    "openai",
    "anthropic",
    "ollama",
    "llamacpp",
    "cli",
    "llm",
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "html2text>=2025.4.15",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-experimental>=0.3.4",
    "langchain-google-community[drive,gmail]>=2.0.7",
    "langchain-openai>=0.3.17",
    "langchain>=0.3.25",
    "langgraph>=0.4.5",
    "pandas>=2.2.3",
    "prompt-toolkit>=3.0.51",
    "pydantic-core>=2.33.2",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "orjson>=3.10.18",
    "typer>=0.15.4",
    "git-python>=1.0.3",
    "pathspec>=0.12.1",
    "numpy",
    "matplotlib>=3.10.3",
    "faker>=37.3.0",
    "rich-pixels>=3.0.1",
    "keyring>=25.6.0",
    "pygithub>=2.6.1",
    "clipman>=3.3.1",
    "pyfiglet>=1.0.2",
    "sixel>=0.2.0",
    "feedparser>=6.0.11",
    "docker>=7.1.0",
    "restrictedpython>=8.0",
    "pyautogui>=0.9.54",
    "pywinctl>=0.4.1",
    "opencv-python>=4.11.0.86",
    "elevenlabs>=1.59.0",
    "pyttsx3>=2.98",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "realtimestt>=0.3.104",
    "kokoro-onnx>=0.3",
    "textual-image>=0.8.2",
    "pyobjc>=11.0.0 ; platform_system == \"Darwin\"",
    "toml>=0.10.2",
    "tomli-w>=1.2.0",
    "redis>=6.1.0",
    "langchain-neo4j>=0.4.0",
    "par-ai-core>=0.3.1",
    "tinify>=1.7.0",
    "langchain-mcp-adapters>=0.1.0",
    "cryptography>=45.0.5",
    "pydantic-settings>=2.0.0",
]
packages = [
    "src/par_gpt",
    "src/sandbox",
]

[project.license]
file = "LICENSE"

[project.urls]
Homepage = "https://github.com/paulrobello/par_gpt"
Documentation = "https://github.com/paulrobello/par_gpt/blob/main/README.md"
Repository = "https://github.com/paulrobello/par_gpt"
Issues = "https://github.com/paulrobello/par_gpt/issues"
Discussions = "https://github.com/paulrobello/par_gpt/discussions"
Wiki = "https://github.com/paulrobello/par_gpt/wiki"

[project.scripts]
par_gpt = "par_gpt.__main__:app"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "build>=1.2.1",
    "pyright>=1.1.379",
    "pre-commit>=3.8.0",
    "ruff>=0.7.0",
    "types-orjson>=3.6.2",
    "docutils-stubs>=0.0.22",
    "pyinstrument>=5.0.0",
]

[tool.hatch.version]
path = "src/par_gpt/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [
    "src/par_gpt",
    "src/sandbox",
]
include = [
    "py.typed",
    "**/*.py",
    "**/*.html",
    "**/*.gif",
    "**/*.jpg",
    "**/*.png",
    "**/*.md",
    "**/*.tcss",
    "**/*.onnx",
    "**/*.json",
    "**/*.yml",
    "**/Dockerfile",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/par_gpt",
    "src/sandbox",
    "LICENSE",
    "README.md",
    "extraction_prompt.md",
    "pyproject.toml",
]
exclude = [
    "*.pyc",
    "__pycache__",
    "*.so",
    "*.dylib",
]
